Update to SpiderMonkey 94beta (#3)
This PR updates to the current beta of SpiderMonkey 94. Previously we were using an arbitrary commit from the Nightly 92 cycle. Switching to beta means that we get more stability, but also get some important improvements that landed in that cycle. Over time, we should switch to stable releases, though.

This PR also improves some aspects of the build script:
- It can now be run from other directories, while still resolving its dependencies correctly
- It continues working correctly when there already is a checkout of the gecko-dev repository, and updates that to the expected revision
- It spends significantly less time on cloning the gecko-dev repository by not unconditionally fetching all tags and branches