Switch to FF 115 with PBL patches applied. (#22)
This PR (i) switches back to the FF 115 base for gecko-dev, and (ii)
applies the PBL patches
(https://github.com/bytecodealliance/gecko-dev/pull/9).

We are reverting to FF 115 because the upgrade to FF 116 seems to have
some issues (or uncover existing issues?); @elliottt is debugging but we
want to decouple this from the initial release of PBL-under-a-flag.

The PBL patches are included, and this commit adds the build flags to
enable them. However, PBL itself is still off by default. The embedder
of the engine needs to set specific configuration options to turn on the
new tier. A subsequent PR to js-compute-runtime will do so when pulling
in this new commit.