Add builds for PBL+weval variant of SpiderMonkey.
This PR adds `release-weval` and `debug-weval` build artifacts,
alongside existing `release` and `debug` artifacts.

It modifies the build script to add the appropriate configuration flags
to the `mozconfig` if invoked as `build-engine.sh {release|debug}
weval`, and adds the two new jobs to CI.

In addition, `download-engine.sh` is modified to support downloading
these variants if invoked as `download-engine.sh {release|debug} weval`.

Names and paths of existing artifacts remain the same, so this is fully
backward compatible. Further work in downstream consumers is needed to
actually make use of the PBL+weval AOT compilation functionality.