Update to Firefox version 140
This includes substantial changes to the build setup, including moving all of the dependency installation out of the shell script and into the CI workflow definition.

Before, we were using the Firefox build system's support for installing toolchains. With this change, we instead install them manually. This means we're not tied to Mozilla's handling of toolchain artifacts, which is explicitly documented as only being supported for current versions of the Firefox codebase.

In combination, these changes lead to faster and easier builds, and make it much easier to do local builds, too.